mc: 简化metrics定义

+++ metrics/src/types/host.rs
@@ -130,5 +130,5 @@ pub fn resource_num_metric(source: &str, namespace: &str, bip: &str, n: u16) {
         metric.zero_num()
     };
 
-    metric += n;
+    metric += n as i64;
 }
